added readbytes client method

+++ maggiefs/dataservice.go
@@ -8,7 +8,7 @@ import (
 
 type DataService interface {
   // read some bytes
-  Read(blk Block, p []byte, pos uint64, length uint64) (err error)
+  Read(blk Block, p []byte, pos uint64, length uint32) (err error)
 
   // write some bytes, extending block if necessary
   // updates generation ID on datanodes before returning
